fix(migration): acceptance_rate is FLOAT, not numeric
Found by spinr-migration-reviewer. drivers.acceptance_rate is FLOAT
(migrations/100_batch_dispatch.sql:52), so declaring it `numeric` in 395's
RETURNS TABLE forced an implicit per-row cast and was inconsistent with
`rating` — the sibling float8 column declared correctly one line above.

Not a crash: float8 -> numeric is an assignment cast, so CREATE FUNCTION and
execution both succeed. Fixing it anyway because the declaration should
mirror the column, and the reviewer's calibration case makes the reason
concrete — the same class of mismatch on destination_mode (boolean returned
as text) would NOT have crashed either, it would have silently disabled the
destination-mode dispatch filter for every driver.

The type test now pins all three columns and asserts no `numeric` remains.

Co-developed with Claude Code (claude-sonnet-5)

+++ backend/migrations/395_dispatch_candidate_drivers_rpc.sql
@@ -74,7 +74,7 @@ RETURNS TABLE(
     lng               double precision,
     rating            double precision,
     is_wav            boolean,
-    acceptance_rate   numeric,
+    acceptance_rate   double precision,
     destination_mode  boolean,
     destination_lat   double precision,
     destination_lng   double precision,
